developed bot play
