+++ README.md
@@ -1,4 +1,4 @@
-# RSSchool NodeJS websocket task template
+# RSSchool NodeJS websocket task
 
 > Static http server and base task packages.
 > By default WebSocket client tries to connect to the 3000 port.
